// examples/export-data/utils/types.ts
export interface Project {
    _id: string | Types.ObjectId;
    clientApiKey: string;
    serverApiKey: string;
    projectName: string;
    privateData: boolean;
    deletionPolicy: number;
    createdAt?: Date;
    updatedAt?: Date;
}
// ...
export enum EventEnvironment {
    DEVELOPMENT = "development",
    STAGING = "staging",
    PRODUCTION = "production"
}